moving python script
